backend/requirements.txt
pandas==2.2.3
supabase==2.28.3
google-genai==1.73.1
python-dotenv==1.0.1
Faker==30.8.2
fastapi==0.115.5
uvicorn[standard]==0.32.1
